+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.16.1](https://github.com/zeon256/satay-rs/compare/satay-codegen-v0.16.0...satay-codegen-v0.16.1) - 2026-08-18
+
+### Fixed
+
+- *(codegen)* shorten runtime serde module paths in
+- *(codegen)* import runtime serde modules in generated types
+
 
 ### Fixed